move sound server to config
