+++ src/recording/audio_recorder.rs
@@ -1,8 +1,6 @@
-use std::fmt::Display;
 use std::fs::create_dir_all;
 use std::path::Path;
 use std::process::Command;
-use std::str::FromStr;
 use std::time::Instant;
 
 use anyhow::anyhow;
@@ -11,40 +9,14 @@ use anyhow::Result;
 use log::debug;
 use regex::Captures;
 use regex::Regex;
-use serde::Deserialize;
-use serde::Serialize;
 use subprocess::Exec;
 use subprocess::Popen;
 
 use crate::config::Opts;
 use crate::config::Service;
+use crate::config::SoundServer;
 use crate::consts::STRIPUTARY_SINK_NAME;
 use crate::recording_session::SessionPath;
-
-#[derive(Copy, Clone, Debug, Default, Deserialize, Serialize, PartialEq)]
-pub enum SoundServer {
-    #[default]
-    Pulseaudio,
-    Pipewire,
-}
-
-impl FromStr for SoundServer {
-    type Err = serde_yaml::Error;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        // This is quite ugly, but ensures that the config file string representation
-        // is the same as in the command line options (which uses the FromStr impl),
-        // without adding any additional dependencies
-        serde_yaml::from_str(s)
-    }
-}
-
-impl Display for SoundServer {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        // Similar to the from_str implementation, this is ugly but consistent.
-        write!(f, "{}", serde_yaml::to_string(self).unwrap())
-    }
-}
 
 pub struct AudioRecorder {
     process: Popen,
