+++ src/recording/mod.rs
@@ -3,4 +3,3 @@ pub mod dbus;
 pub mod dbus_event;
 pub mod recorder;
 pub use audio_recorder::AudioRecorder;
-pub use audio_recorder::SoundServer;
